Refactor: Split useDesignState hook into smaller files
This commit refactors the `useDesignState` hook into smaller, more manageable files to improve code organization and readability.
